New Component project structure New Samples Project Folder Refactored DS1307 Sample which uses RasperrySharp.Components Namespace Header for the Source Files including a template header Some enhancements for the Readme Nuget Update for NLog and UnitsNet in the Winforms Demo
