+++ Samples/Advanced/Demo_Winforms/Buses/I2CBus.cs
