+++ Samples/Advanced/Demo_Winforms/Buses/I2CConnectorDriver.cs
