+++ Samples/Advanced/Demo_Winforms/Main.resx
