+++ Samples/Advanced/Demo_Winforms/Modules/DS18B20Service.cs
