+++ Samples/Advanced/Demo_Winforms/Properties/AssemblyInfo.cs
