+++ Samples/Advanced/Demo_Winforms/Properties/Resources.resx
