+++ Samples/Advanced/Demo_Winforms/Properties/Settings.settings
