+++ Samples/Advanced/Demo_Winforms/Utilities/CommonHelper.cs
