+++ Samples/Advanced/Demo_Winforms/packages.config
@@ -1,0 +1,5 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="NLog" version="4.5.2" targetFramework="net462" />
+  <package id="UnitsNet" version="3.94.0" targetFramework="net462" />
+</packages>
