+++ Samples/Clocks/Ds1307 (I2C)/App.config
