+++ Samples/Gpio/WatchPin/Properties/AssemblyInfo.cs
