+++ Samples/Gpio/WatchPin/WatchPin.csproj
