+++ Samples/Monitors/LTC2943/App.config
