+++ Samples/Monitors/LTC2943/LTC2943.csproj
