+++ Samples/Monitors/LTC2943/Program.cs
