+++ Samples/Monitors/LTC2943/Properties/AssemblyInfo.cs
