+++ Samples/Sensors/Temperature/DS18B20/App.config
