+++ Samples/Sensors/Temperature/DS18B20/DS18B20.csproj
