+++ Samples/Sensors/Temperature/DS18B20/Properties/AssemblyInfo.cs
