--- Tests/Demo/packages.config
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="NLog" version="4.4.12" targetFramework="net462" />
-  <package id="UnitsNet" version="3.89.0" targetFramework="net462" />
-</packages>
